+++ pkg/grpc_server/server/h_transaction_test.go
@@ -165,7 +165,7 @@ func TestHandlerTransaction(t *testing.T) {
 
 		currTx, err = intMaxTypes.NewTx(
 			&transferRoot,
-			uint64(nonce),
+			nonce,
 		)
 		assert.NoError(t, err)
 	}
